// src/app/Models/emailConfig.ts
export class emailConfig {
  bcc: string;
  buId: string;
  buName: string;
  cc: string;
  createdBy: string;
  createdByName: string;
  createdDateTime: string;
  id: string;
  message: string;
  purpose: string;
  status: string;
  subject: string;
  to: string;
  updatedBy: string;
  updatedByName: string;
  updatedDateTime: string;
  keyword: string;

  searchBuId: string;
  email: string;
  toDateTime: string;
  fromDateTime: string;
}